profe, esto lo hice el dia despues del parcial para entender bien como hacer la llamada a la API y la conversion de los elementos que llegaban desde la misma, el parcial que hice ayer esta en el master

+++ src/main/java/com/example/parcial/martin/caminero/model/Currency.java
@@ -19,5 +19,5 @@ public class Currency {
 
     private CurrencyType currency;
 
-    private int total;
+    private float total;
 }
